+++ densEstAI/core/__init__.py
@@ -1,4 +1,4 @@
-from .video.streamer import VideoStreamer
-from .video.streamer_func import start_stream
-from .video.streamer_thread import main 
-from .html_generater import HtmlGenerator
+from .stream.streamer_func import start_stream
+from .stream.streamer_thread import main 
+from .yolo.preprocessor import PreProcessor 
+from .yolo.yolo_manager import YoloManager
